declare routes
Inside of your App.js file import { Route } from 'react-router-dom';
This is where we're going to declare and specify our router.
Create 3 <Route /> setting their path prop equal to /, /about, /contact with their respective components.
Be sure to incluce the exact prop on the root component for / to make sure that it's rendering the exact component and not all the other components.
You should now be able to type /, /about, /contact after localhost:PORT/